+++ layout/generic/nsTextFrameThebes.cpp
@@ -4865,7 +4865,7 @@ ComputeSelectionUnderlineHeight(nsPresContext* aPresContext,
         aPresContext->AppUnitsToDevPixels(nsStyleFont(aPresContext).mFont.size);
       gfxFloat fontSize = std::min(gfxFloat(defaultFontSize),
                                  aFontMetrics.emHeight);
-      fontSize = std::max(fontSize, 1.0);
+      fontSize = std::max(fontSize, (double)1.0);
       return ceil(fontSize / 20);
     }
     default:
